Intern (#7)
A great increase in compile speed (from minutes to 2 seconds) for extremely large queries of over hundreds of thousands of characters (800 000).  
This has been achieved by:

  - using smart constructors instead of out of band simplification, for patterns and funcs.
  - including a hash that is calculated only once in each pattern and func, for faster comparison

An `intern` package was created as an alternative to the `ast` and `interp` packages.  It throws away all syntax and houses the hashes for patterns.  The hope is that this will also result in less pointers for the garbage collector to check.

Funcs had some major changes.
They now require more methods: Compare, Hash, HasVariable, String
Trimming is done in the constructor instead of magical reflect.
Compare should first check the hash and then try to compare with its own type and only as a last resort create a string representation to do a comparison with an unknown func which has the same hash.  The string creation was really slow for large functions.
The reflection magic in sprint has also been removed in favor of a String method which is implemented by the function itself.
The combinator no longer using sprint, so this String method does not need to parsable anymore, but it is still preferable.
Init methods have also been abandoned in favor of this logic happening the constructor.  This constructor is now also responsible for simplification.
Registering functions now no longer requires a pointer to a struct, but rather a constructor function.

Lastly the mem package has also been greatly simplified and more errors are returned rather than panicing the the nameexpr package.